app/requirements.txt
flask
SQLAlchemy
Flask-SQLAlchemy
Werkzeug
flask-marshmallow
requests
sympy
python-socketio
python-socketio[client]